requirements.txt
fastapi
uvicorn
duckdb
openai
python-dotenv
pydantic
pandas